refactor: 플랜 작성자 검증 로직 변경 (#15)

+++ src/main/kotlin/kr/wooco/woocobe/plan/domain/usecase/DeletePlanUseCase.kt
@@ -16,11 +16,12 @@ class DeletePlanUseCase(
 ) : UseCase<DeletePlanInput, Unit> {
     @Transactional
     override fun execute(input: DeletePlanInput) {
-        val findPlan = planStorageGateway.getById(input.planId)
+        val plan = planStorageGateway.getById(input.planId)
             ?: throw RuntimeException("Not exists plan")
 
-        findPlan.takeIf { it.isWriter(input.userId) }
-            ?: throw RuntimeException("Invalid plan writer")
+        when {
+            plan.isWriter(input.userId).not() -> throw RuntimeException()
+        }
 
         planStorageGateway.deleteById(input.planId)
     }
